uniformchat/requirements.txt
asgiref==3.5.0
attrs==19.3.0
Automat==0.8.0
backcall==0.1.0
backports.zoneinfo==0.2.1
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
Click==7.0
cloud-init==21.4
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==36.0.2
cupshelpers==1.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
Deprecated==1.2.13
distlib==0.3.4
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.0.3
django-cors-headers==3.11.0
django-debug-toolbar==3.2.4
django-oauth-toolkit==1.7.1
entrypoints==0.3
filelock==3.6.0
gunicorn==20.1.0
httplib2==0.14.0
hyperlink==19.0.0
idna==3.3
importlib-metadata==1.5.0
incremental==16.10.1
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.15.2
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
jwcrypto==1.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
macaroonbakery==1.3.1
MarkupSafe==1.1.0
more-itertools==4.2.0
netifaces==0.10.4
oauthlib==3.2.0
parso==0.5.2
pexpect==4.6.0
pickleshare==0.7.5
pipenv==2022.3.28
platformdirs==2.5.1
prompt-toolkit==2.0.10
protobuf==3.6.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycparser==2.21
pycups==1.9.73
Pygments==2.3.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
pytz==2019.3
PyYAML==5.3.1
requests==2.27.1
requests-unixsocket==0.2.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.16.0
slack-sdk==3.15.2
sos==4.1
sqlparse==0.4.2
ssh-import-id==5.10
systemd-python==234
traitlets==4.3.3
Twisted==18.9.0
tzdata==2022.1
ubuntu-advantage-tools==27.4
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.9
virtualenv==20.14.0
virtualenv-clone==0.5.7
wadllib==1.3.3
wcwidth==0.1.8
wrapt==1.14.0
zipp==1.0.0
zope.interface==4.7.1